add:認証機能の実装

+++ app/Http/Controllers/CommentController.php
@@ -2,8 +2,10 @@
 
 namespace App\Http\Controllers;
 
+use Auth;
 use App\Models\Comment;
 use App\Models\Snack;
+use App\Models\User;
 use Illuminate\Http\Request;
 use App\Http\Requests\CommentRequest;
 use Illuminate\Pagination\Paginator;
@@ -63,10 +65,17 @@ class CommentController extends Controller
      */
     public function show(Comment $comment)
     {
-        return view("comments.show")->with([
-            'comment' => $comment,
-            'bookmark_list' => auth()->user()->comments()->get()
-        ]);
+        if( Auth::check()){
+            return view("comments.show")->with([
+                'comment' => $comment,
+                'bookmark_list' => auth()->user()->comments()->get()
+            ]);
+        } else {
+            return view("comments.show")->with([
+                'comment' => $comment
+            ]);
+        }
+        // Auth::check()で、ログイン済みの時の処理ができる.
     }
 
     /**
@@ -77,9 +86,14 @@ class CommentController extends Controller
      */
     public function edit(Comment $comment)
     {
-        return view('comments.edit')->with([
-            'comment' => $comment
-        ]);
+        $user = auth()->user();
+        if($user->can('update', $comment)){
+            return view('comments.edit')->with([
+                'comment' => $comment
+            ]);
+        } else {
+            return redirect('/comments/' . $comment->id);
+        }
     }
 
     /**
@@ -107,12 +121,18 @@ class CommentController extends Controller
      */
     public function delete(Comment $comment)
     {
-        $comment->delete();
-        return redirect('/snacks/' . $comment->snack->id);
+        $user = auth()->user();
+        if($user->can('delete', $comment)){
+            $comment->delete();
+            return redirect('/snacks/' . $comment->snack->id);
+        } else {
+            return redirect('/comments/' . $comment->id);
+        }
     }
     
     public function bookmarked()
     {
+        
         $comment = auth()->user()->comments()->orderBy('created_at', 'desc')->paginate(10);
         return view('comments.bookmarked')->with([
             'comments' => $comment,
