+++ resources/views/snacks/index.blade.php
@@ -18,18 +18,24 @@
                         <a href="/snacks/{{ $snack->id }}">{{ $snack->name }}</a>
                     </h2>
                     <p class='overview'>{{ $snack->overview }}</p>
-                    <form action="/snacks/{{ $snack->id }}" id="form_{{ $snack->id }}" method="post" style="display:inline">
-                        @csrf
-                        @method('DELETE')
-                        <button type="submit", onclick="return deleteSnack()" style='padding: 10px 0 0 0;'>[削除]</button> 
-                    </form>
+                    @if(Auth::id() == implode(config('app.admin')))
+                        <form action="/snacks/{{ $snack->id }}" id="form_{{ $snack->id }}" method="post" style="display:inline">
+                            @csrf
+                            @method('DELETE')
+                            <button type="submit", onclick="return deleteSnack()" style='padding: 10px 0 0 0;'>[削除]</button> 
+                        </form>
+                    @endif
                 </div>
             @endforeach
         </div>
         <div class='paginate' style='padding: 0 0 0 70px;'>
             {{ $snacks->links() }}
         </div>
-        <a href='/snacks/create' style='padding: 0 0 0 50px;'>[お菓子の追加]</a>
+        
+        @if(Auth::id() == implode(config('app.admin')))
+            <a href='/snacks/create' style='padding: 0 0 0 50px;'>[お菓子の追加]</a>
+        @endif
+        {{-- divで<a>タグを囲む --}}
         
         <script>
             function deleteSnack() {
